+++ CHANGELOG.md
@@ -78,6 +78,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ExtrusionSurface3D: linesegment2d_to_3d.
 - BSplineSurface3D: ban useless attr in serialization
 - BSplineCurve: simplify
+- SphericalSurface3D: contour3d_to_2d
 
 
 ### Refactor